+++ src/test/groovy/edu/ie3/datamodel/io/source/influxdb/InfluxDbWeatherSourceIT.groovy
@@ -24,10 +24,9 @@ import spock.lang.Specification
 class InfluxDbWeatherSourceIT extends Specification implements WeatherSourceTestHelper {
 
 	@Shared
-	InfluxDBContainer influxDbContainer = new InfluxDBContainer()
+	InfluxDBContainer influxDbContainer = new InfluxDBContainer("latest")
 	.withAuthEnabled(false)
 	.withDatabase("test_weather")
-	.withExposedPorts(8086)
 
 	@Shared
 	InfluxDbWeatherSource source